+++ src/main/java/org/sopt/domain/post/exception/PostException.java
@@ -1,6 +1,6 @@
 package org.sopt.domain.post.exception;
 
-import org.sopt.global.CustomException;
+import org.sopt.global.exception.CustomException;
 
 public class PostException extends CustomException {
