+++ src/main/java/org/sopt/domain/user/exception/UserErrorCode.java
@@ -1,6 +1,6 @@
 package org.sopt.domain.user.exception;
 
-import org.sopt.global.code.ErrorCode;
+import org.sopt.global.api.code.ErrorCode;
 import org.springframework.http.HttpStatus;
 
 public enum UserErrorCode implements ErrorCode {
