+++ src/main/java/org/sopt/global/exception/GlobalExceptionHandler.java
@@ -1,6 +1,7 @@
-package org.sopt.global;
+package org.sopt.global.exception;
 
-import org.sopt.global.code.CommonErrorCode;
+import org.sopt.global.api.CommonResponse;
+import org.sopt.global.api.code.CommonErrorCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -11,25 +12,25 @@ public class GlobalExceptionHandler {
 
     // CustomException을 상속한 모든 예외 처리
     @ExceptionHandler(CustomException.class)
-    public ResponseEntity<ApiResponse<Void>> handleCustomException(CustomException e) {
+    public ResponseEntity<CommonResponse<Void>> handleCustomException(CustomException e) {
         return ResponseEntity
                 .status(e.getErrorCode().getHttpStatus())
-                .body(ApiResponse.error(e.getErrorCode()));
+                .body(CommonResponse.error(e.getErrorCode()));
     }
 
     // JSON 파싱 실패
     @ExceptionHandler(HttpMessageNotReadableException.class)
-    public ResponseEntity<ApiResponse<Void>> handleHttpMessageNotReadableException(HttpMessageNotReadableException e) {
+    public ResponseEntity<CommonResponse<Void>> handleHttpMessageNotReadableException(HttpMessageNotReadableException e) {
         return ResponseEntity
                 .status(CommonErrorCode.BAD_REQUEST.getHttpStatus())
-                .body(ApiResponse.error(CommonErrorCode.BAD_REQUEST));
+                .body(CommonResponse.error(CommonErrorCode.BAD_REQUEST));
     }
 
     // 그 외 예외 처리
     @ExceptionHandler(Exception.class)
-    public ResponseEntity<ApiResponse<Void>> handleException(Exception e) {
+    public ResponseEntity<CommonResponse<Void>> handleException(Exception e) {
         return ResponseEntity
                 .status(CommonErrorCode.INTERNAL_SERVER_ERROR.getHttpStatus())
-                .body(ApiResponse.error(CommonErrorCode.INTERNAL_SERVER_ERROR));
+                .body(CommonResponse.error(CommonErrorCode.INTERNAL_SERVER_ERROR));
     }
 }
